feat(client): VP進捗カードを共有化し応援画面に適用、次の称号名を非表示に
- 累計VPと次の称号までの進捗カードを VpSummaryCard として共有化し、
  業績画面と応援画面の両方で同じデザインを使用
- 次の称号名と色は到達時の楽しみのため伏せる
- 役割を統合した VPProgressWidget を削除し、テストを追従
